requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
google-genai
numpy==2.1.1
pypdf==5.0.1
python-docx==1.1.2